requirements.txt
flask
langchain
langchain-groq
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
flask
langchain
langchain-groq
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
pymupdf4llm
python-dotenv
beautifulsoup4
requests
pymongo
dnspython
gunicorn